Refactored the healing logic
