+++ packages/wdio-browserstack-service/src/constants.ts
@@ -70,9 +70,14 @@ export const LOG_KIND_USAGE_MAP = {
 
 export const FUNNEL_INSTRUMENTATION_URL = 'https://api.browserstack.com/sdk/v1/event'
 
+export const SUPPORTED_BROWSERS_FOR_AI = ['chrome', 'edge', 'firefox']
+
 export const TCG_URL = 'https://tcg.browserstack.com'
 
-export const SUPPORTED_BROWSERS_FOR_AI = ['chrome', 'edge', 'firefox']
+export const TCG_INFO = {
+    tcgRegion: 'use',
+    tcgUrl: TCG_URL,
+}
 
 // Env variables - Define all the env variable constants over here
